refactor(metrics): rename metrics_base.h to base.h
