+++ include/endstone/metrics/base.h
